refactor: extract reusable template agent flow
