+++ src/xtalk/llm_agent/__init__.py
@@ -1,6 +1,7 @@
 from .interfaces import Agent
 from .dummy import DummyAgent
 from .default import DefaultAgent
+from .template import MutableToolProvider, TemplateAgent
 from .runtime import (
     AgentRequest,
     AgentRuntime,
@@ -10,7 +11,9 @@ from .runtime import (
     PromptBuilder,
     ScenarioSpec,
     TextChunkEvent,
+    ToolCallEvent,
     ToolProvider,
+    ToolResultEvent,
     TurnContext,
     TurnHook,
 )
@@ -19,15 +22,19 @@ __all__ = [
     "Agent",
     "DummyAgent",
     "DefaultAgent",
+    "TemplateAgent",
     "AgentRequest",
     "AgentRuntime",
     "AgentSession",
     "ContextAdapter",
+    "MutableToolProvider",
     "OutputPolicy",
     "PromptBuilder",
     "ScenarioSpec",
     "TextChunkEvent",
+    "ToolCallEvent",
     "ToolProvider",
+    "ToolResultEvent",
     "TurnContext",
     "TurnHook",
 ]
